Implement start_ssl_listener, to complete ssl acceptor implementation
- Implement ssl_connection_upgrade/2 callback, to use in upgrading
a new tcp_connection to an ssl_connection.
- ssl_connection_upgrade then calls out to start_ssl_client to
setup start the ampq session. Right now, this is just a dummy
session, but when the async_recv patches are done, it will become
a full valid session.
- Abstract TCP options into a macro to help reuse.
- Currently, I'm hardcoding location for cacerts, certfile and private
key file. Before this is finalized, config parameters will be agreed,
and used to supply this information.