Added means to alter all queues and switch to disk_only mode in the disk queue.
rabbit_queue_mode_manager:change_memory_usage(undef, true).

this will first ask all queues to switch from mixed to disk mode, and will on a 2nd call, ask the disk queue to switch to disk only mode.

rabbit_queue_mode_manager:change_memory_usage(undef, false).

moves the other way.

This all works, eg set MulticastMain pushing in messages and switch modes, and it's fine.

One immediate problem is that as soon as everything becomes disk only, the performance suffers, so as a result messages build up. This is as expected. Then, going back to the middle mode (i.e. disk queue in ram_disk mode and queues in disk mode), the switch in the disk queue eats up a lot of memory. I suspect this is the effect of converting the mnesia table from disc_only_copies to disc_copies when there are 40k+ messages in there (one row per message). As a result, this conversion on its own is very dangerous to make. It might be more sensible to use the "weird" mode, where the queues are in mixed mode and the disk queue is in disk_only mode so as to try and get the queues to drain as fast as possible, reducing the size of the mnesia table so that when it is finally converted back, it's small.

More experimentation is needed.

I'll hook the above commands into rabbitmqctl soon.