Right, whilst tx_commits were being coalesced, this was not done properly protect non-tx publishes from unnecessary syncs. As a result, performance, of say:
rabbitmq-java-client/build/dist$ sh runjava.sh com.rabbitmq.examples.MulticastMain -r 3500 -s 12345 -f persistent -z 20

was horrible when compared to bug19662 - about 5 times slower.

Careful consideration of when we want to set the timer, when to cancel, and when the timeouts should be set has led to this revision which maintains the tx performance and restores non-tx performance.