Omit runlevel list in the LSB Default-{Start,Stop} comment
Neither our Fedora/RH or Debian packages actually depend on the value
in this comment - the list of runlevels to install the service into
comes from somewhere else in both cases.  But Fedora guidelines say
that "Only services which are really required for a vital system
should define runlevels here".  So don't.