Made the disk queue start up in the same mode it was last running in.
This is slightly grim because I have to store some values in the mnesia table which then have to survive all the start up logic, so there are a couple of annoying 1-line changes elsewhere. However, it does indeed work.

There was also one bool() -> boolean() fix in the memory_manager.