initialise msg_store ref counts via a generator This is more flexible than the previous ref_count function, allowing the ref counts to be obtained w/o consuming any memory at the supplying end in a variety of scenarios.
We use the dq_msg_loc ets/dets table to store the ref counts. That
table is later updated with the full details of the messages (their
file and position, etc). At the end we prune any entries that have a
ref count but no associated file - i.e. the referenced message
couldn't be found on disk.

This change should also fix the "All replicas on diskfull nodes are
not active yet" error observed in bug 21530 since we no longer need
the indices on the rabbit_disk_queue mnesia table which we identified
as the most likely cause of that error.