drop disk_queue/msg_store disk_only mode and all the mode switching and memory management logic that goes with it.
The 2G limitition of dets make the disk_only mode not worthwhile.

In the process I refactored the msg_location access in msg_store
s.t. it shouldn't be much effort to plug in a different index store in
the future.

Also some minor tweaks and tidying up here and there.