Do the queue deletion synchronously.  Instead of replying immediately using gen_server2:call, then doing the work, we include the reply in the return value for gen_server2 to pass on.  It happens that gen_server2 will terminate the process, on getting stop, before it sends the reply on -- just what we want here.
NB: here we translate between the return value of handle_ch_down/2,
which was intended to be the return value for handle_call, and an
actual return value.  Slightly nicer would be for handle_ch_down to
return the state and (say) 'ok' or 'stop', to make its contract more
explicit.