tweak tx API on amqqueue commit_all and rollback_all take the channel pid as an additional arg. This brings these functions in line with deliver and ack, which also take both a txn and ch_pid.
In the queue process this saves us some book keeping.

Also, in the queue process we need to clear the txn field in the ch
record on commit/rollback, since otherwise a subsequent channel 'DOWN'
results in some suprising and unnecessary, though perfectly safe,
control flow.

Finally, there is no need to check that a commit relates to a channel
the queue process knows about - this is always guaranteed to be the
case.

All these changes were cherry-picked from the bug21673 branch and
bring default closer to that branch.