Handle quoted node names by parsing with erl_scan:string
If the node name contains characters which cannot be used in a simple
Erlang atom (e.g. a hyphen), then pid_to_string will quote it;
e.g. <'rabbit@centos5-64'.51.0>.  If in string_to_pid we try to
convert the node name part back to an atom with list_to_atom, the
quotes will be included in the resulting atom
(e.g. '\'rabbit@centos5-64\'') and so the two functions will not be
inverses.  So use erl_scan to parse the node name instead, to recover
an atom that exactly matches the original.