Multi-protocol broker. Let joy be unconfined.
The resultant broker passes the entire 0-8 test suite on default. Unfortunately we can't run the full 0-9-1 test suite as that ends up trying to rebuild part of the broker which fails due to rabbitmq-codegen being on the wrong branch. However the Java functional tests all pass bar one; the exception is UnexpectedFrames.testMissingHeader which fails due to different error codes used on default. This will be fixed by bug 22886.