Remove IFS setting on the su command
The script has lots of IFS vulnerabilities before it reaches the su
command.  If we really want to guard against IFS, we should set it at
the top.  But how many shell scripts have you ever seen that do that?
If the user sets IFS, and breaks things, it seem reasonable to regard
that as their problem.  There is no security issue here because you
should already be root or the rabbitmq user to run the script.