homemade opt parser
It's either this or ~400 lines of code from github under a
non-standard license.

The idea here is that you specify options with tuples like:
  - {flag, <flag string>}, or
  - {option, <option string>, <default value>}.

Then, you let get_options separate the flags and options from the
arguments.

The spec for get_options is:

get_options/2 :: ([{flag, string()} | {option, string(), any()}],
                  [string()])
                 -> {[string()], [{string(), any()}]}

For instance:
   get_options([{flag, "-q"}, {option, "-p", "/"}],
               ["set_permissions","-p","/","guest", "-q",".*",".*",".*"])
== {["set_permissions","guest",".*",".*",".*"],
     [{"-q",true},{"-p","/"}]}