new logic
Message sequence numbers are assigned uniquely when the basic.publish
is received.  If confirm is not enabled, this is undefined.

Rabbit_channel gains a new API method: confirm.  This schedules an ack
for the given sequence number.  The idea is that repeated acks are
allowed at this point and de-duplication is done by the channel.  This
way, in order to confirm a message, another module just needs to
know 1) the channel and 2) the message sequence number.

No de-duplication is done at this point.