fixed an ever-expanding dictionary
The problem with the current design is that the lowest layer that knows
about confirmations is amqqueue_process.  This means that
amqqueue_process, variable_queue, msg_store and queue_index need to do
all the bookkeeping associated with confirmations even if they're not
enabled.  Mind you, this doesn't slow things down noticeably.