fixed message leak caused by confirms
I've added variable_queue:init/4.  It calls variable_queue:init/3 and
the fourth parameter is a tuple of 3 functions: the function to call
when the indices of some guids hit the disk, the function to call when
some guids hit the disk and the function to call when some guids have
been ack'd by the consumer.  Rabbit_test supplies nops rather than the
usual casts, so the superfluous messages aren't generated.