backport upstream change to way print_event/3 is called The code previously relied on the ancient {ModuleName,FunName} syntax to denote functions. That is bad for dialyzer and also forces an unnecessary export.
The underlying code in sys:handle_debug hasn't actually changed - I've
checked back as far as R12B-3 and it's always been invoking the fun
with Fun(...) and is not doing any pattern matching on it. Also, the
R12B-3 docs says that the fun is "fun(FuncState,Event,ProcState) ->
done | NewFuncState", i.e. there is no hint that it should be passed
as a {ModuleName,FunName} pair.