add stats for confirms
To recap, a published message is confirmed by the channel.  A message
is confirmed only after all the queues it was published to confirm it.

With the current change, the emitted stats look like this:

        {channel_exchange_stats,
            [{{resource,<<"/">>,exchange,<<"direct">>},
              [{confirm,545},{publish,545}]}]},
        {channel_queue_exchange_stats,
            [{{<0.204.0>,{resource,<<"/">>,exchange,<<"direct">>}},
              [{confirm,545},{publish,545}]},
             {{<0.195.0>,{resource,<<"/">>,exchange,<<"direct">>}},
              [{confirm,545},{publish,545}]}]}]

The confirm field in channel_exchange_stats represents the number of
messages sent to that exchange that have also been confirmed.  If the
exchanged routed the message to different queues, this number is only
increased when all queues have confirmed the message.  If the message
was unroutable or was routed to 0 queues, this number is still
increased.  This is the number of basic.confirms sent back to
publisher.

The confirm field in channel_queue_exchange_stats represents the
number of messages confirmed by that queue (but not necessarily
confirmed by the channel).

In channel_exchange_stats, if the number of confirms lags behind the
number of publishes, one of the queues is not confirming messages in a
timely fashion.