enable disc-less nodes
The main trouble with this is
mnesia:change_config(extra_db_nodes,...).  If the current node doesn't
have a schema with all the tables created, it will get it from one of
the nodes it connects to.  This is a problem if we're a disc node and
they're a ram node or vice versa.  So, after we get the tables from
the other node and wait_for_remote_tables, we assert that their
storage location is correct and change it if not.  If the other node
is disc and we are ram, this unfortunately means that the tables are
first written to disc and then converted to ram, but since this is a
one-off operation during startup, it shouldn't be a problem.