test that disc nodes are actually disc nodes
The tests used to be slightly wrong.  They'd expect:
  rabbitmqctl cluster
to create a cluster with a single ram node.  It didn't, instead, it
created a cluster with a single disc node.  Since we weren't actually
checking the resulting node type, everything went along fine.

Assert_tables_copy_type is complicated by all the cases it has to
handle, namely converting disc -> ram and ram -> disc.  For disc ->
ram, we first need to convert all the tables to ram, then, we need to
convert the schema (converting it before fails with "Disc resident
tables").  For ram -> disc, we first need to convert the schema,
otherwise, all the table conversions will fail with 'has_no_disc'.

Regarding an earlier commit, using mensia:system_info(use_dir) to
check if we have a disc node is wrong because we create the directory
for the message_store anyway.