remove two resets introduced earlier
We don't need them anymore.

BTW, when generating the backup db directory name, we used to just
append the date (including minutes, seconds) to the current directory
name.  But since we do an extra move_db/0 now (when converting from
disc to ram), some of the tests were fast enough to try to backup the
database twice in the same second.  Needless to say, this would cause
an error.  So, now we generate the new name as before, but if it's
already used, we wait 1s and try again.