reset before changing node type
Matthias points out that the only way for mnesia:change_config to
start creating tables is for the node to re-join the cluster (and get
its old table definitions from other nodes).  Instead of fixing the
tables after that, we now reset inside the cluster command.  Note that
in very old rabbit versions, we'd just prevent this by having the user
reset manually.