always reset and preemptively leave clusters
Always backup and reset during clustering if the node type has
changed.  Worst case, it ensures that we actually have an empty mnesia
dir with new ram nodes.

Also, preemptively leave a cluster before joining it.  Suppose we had
a two-node cluster, the first node goes down, the second hard resets,
the first node comes back up, the second node tries to rejoin the
cluster with a different type.  Since it hard-reset, it doesn't know
that it used to be part of the cluster, and the other node is unaware
that our node is supposed to have left the cluster.  So, when
clustering, we always try to leave a cluster before joining it.

In leave_cluster/2, I added {aborted, {node_not_running, _}} to the
"not error" returns, because it looks similar to {badrpc, nodedown},
which was already there.  This may be wrong.