only allow the rabbit error logger to touch the files and append rather than overwrite
If the standard error logger is initialised with the correct file
names, the files will be overwritten.  So, we no longer tell Erlang
the location of the files at boot.  It will automatically register an
error_logger_tty_h, which we then unregister after setting up our own
loggers (you can't swap a file logger for a tty logger).

Still left todo:
  - error loggers when stopping rabbit,
  - windows rabbitmq-server.bat,
  - the other head of ensure_working_log_handlers,
  - it doesn't really make sense to swap out the non exitent
    error_logger_file_h.