if the worker_pool is not running, run the submitted job in the current process
The only time when this happens is when using the file2 and filelib2 functions
during prelaunch and prepare, which should be fine.

The alternative is starting worker_pool before it's needed, but the logic of
that ends up being quite complicated, as we also have to stop it before
starting rabbit fully and it may already be running.