Buffer transfers that we cannot send because of the session window being closed. We cannot in general stop the queue from sending new messages, especially when 1. the link credit (denominated in messages) exceeds the session's capacity for transfers (denominated in frames) 2. acking is switched off, in which case basic.qos is not in use anyway
Therefore, we try to set basic.qos to have *some* control, and ack after a message has actually been fully transferred (or if we are sending unsettled, when the outcome comes back from the client).
We also keep a queue of transfers yet to be sent (grouped by message, so we know when a message has been sent in full).