add special case for the only error I could find
There are two things that may break network-wise when starting rabbit: epmd
may fail to start for some reason, net_kernel may fail to start for some
reason.  We're now starting epmd manually, because net_kernel needs it; also,
there doesn't seem to be a way to start it from Erlang, so we have to start it
from the shell scripts.  Of course, running it in daemon mode hides any
errors it may encounter completely; i.e. there's no way to tell if "epmd
-daemon" actually started the daemon.

There isn't any documentation for what errors net_kernel:start/1 may return,
so we print a vague error message and exit in case of an error.  There's also
a bit of special handling for the case in which epmd didn't start (detected
because something deep down in Erlang fails to start).