Split guid/0 in guid/0 and the faster fast_guid/0.
The first function works like the old one (hashes the {serial, counter} each
time.

The second function (fast_guid/0) hashes the serial once, and the XORs it with
the counter each time a new guid is needed. Thus, it is more predictable but
should not produce less unique guids.

If the fast function is used when generating ids for messages, it leads to a
considerable speedup when testing with BroadcastMain (from 67k msg/s to
75k msg/s).