move ssl upgrade from acceptor to reader ...thus allowing further connections to be accepted while the upgrade takes place.
The reader & processor startup needed to be re-arranged. The processor
needs to know about the socket, but that now doesn't become available
until the reader has upgraded it. Hence we can no longer start the
processor at the same time as the reader. Instead we get the reader to
start the processor.

We also need to change the reader's restart type from 'temporary' to
'intrinsic'. The former was always a bit dodgy since it meant a reader
exit would not terminate the supervisor, thus leaking a process. In
practice that was generally prevented due to the reader containing a
'try ... after' that would terminate the processor, which is an
'intrinsic' process and hence would take down the supervisor. But the
reader can exit before entering the 'try ... after', e.g. if
rabbit_net:peername fails (rare, but certainly possible in cases where
a client connects and immediately disconnects), and, with the changes
in this bug, when the ssl upgrade fails.