Syncing instead of flushing.
This stops needs_timeout and timeout from going in a loop. However, since
rabbit_queue_index:needs_sync works looking at unsynced_msg_ids, this doesn't
preserving acks when the broker is terminated abruptedly, since the timeout
won't be triggered (needs_sync will return false).

If needs_sync is defined inspecting the write_buffer if the file_handle_cache,
needs_timeout and timeout go in a loop, for reason that are not clear to me
now.