Add join_cluster, chang the way cluster config works
The arguments of `join_cluster' are a set of nodes to use to discover the disc
nodes present in the cluster. The old `cluster' command now emits a warning
and is actually `join_cluster'. I still need to make the configurations update
when a new disc node joins the cluster.

The node's cluster config now stores the discovered nodes and whether
the current node is on disc or not. This is less hackish then the previous
solutions (the node is on disc if it's included in the list of cluster nodes
in the config), and makes some code in `rabbit_upgrade' nicer. Note that this
might change in the future anyways, since it doesn't make much sense for a disc
node to have that kind of config.

Things that I'm not sure of:
  * What does `force' mean in this context. Right now I left the
    `force_cluster' command in place, and it calls
    `rabbit_mnesia:force_cluster' as before. The only difference between force
    and not force is that when `force = true' and the node cannot discover the
    disc nodes, it simply uses the provided nodes. I'm not sure this is
    desirable, but since `force' means "make it work even if the nodes are
    offline" (at least in this context), I think it's the only way out. The
    differences between force/non force in `rabbit_mnesia:init_db/4' remain the
    same as before.
  * What do we do with old configurations. Right now I simply check how
    they're formed and convert them to the new format, but I'm not sure if it's
    better to update the file or not. Again this might change since in the
    future I'll have to set up a way to update the config file anyways.