change function to find out about cluster nodes
See comment above `check_mnesia_running/1' in `rabbit_mnesia'. In short,
the function to find out about all/running nodes in the cluster do not
work if mnesia is down and the node is a ram node, while we assumed
that they would always work. Moreover, the functions to get the cluster
disc nodes require mnesia to be up.