simplify `rabbit_misc:get_options/4'
Incidentally, it is also much more efficient (not that it matters).

I will probably rename it to something like `parse_arguments'.

`CommandOpts' and `GlobalOpts' could be one, but I think that `GlobalOpts'
is really useful since adding the global options to each single command
is a pain. We could, obviously, put them together in a tuple, but then
it's not that different from two separate arguments.