moved the status file handling to `rabbit_node_monitor', status from mnesia when online
I'm much happier with this configuration, with the exception of one
problem: `rabbit_mnesia:cluster_status' will happily return the data from
mnesia whenever mnesia is online, but that gives wrong results when the
node is a RAM node and it just got online. I have some ideas to improve this.