fix `remove_node/1' - do not remove nodes from offline nodes, unless we're "safe"
We wan to remove nodes from offline nodes only when all the cluster nodes
are offline, and when the node we are removing from was either the last
to go down or the second after the node we are removing. However we cannot
know that for sure. For this reason rabbit will refuse to remove from offline
nodes unless the flag `--offline' is passed.

There is more information in comments around `rabbit_mnesia:remove_node/2.