revert last change (reset when changing node from disc to ram)
mnesia, mnesia....

For some reason, this approach does not in some cases.  For example, starting
with nodes A, B, and C clustered and all disc, the following sequence of
commands causes an error:

  $ rabbitmqctl -n A stop_app
  $ rabbitmqctl -n B stop_app
  $ rabbitmqctl -n B change_cluster_node_type ram # reset happens here
  $ rabbitmqctl -n B start_app
  $ rabbitmqctl -n A start_app
    error happens here:
    {error, {cannot_start_application,mnesia,
             {shutdown,{mnesia_sup,start,[normal,[]]}}}}

I have no idea why, but it's probably connected to the fact that we can't do
those changes when disc replicas are offline, and probably also related that
operations like leaving the cluster require all the cluster node to be online.

At this point I'd give up and put this limitation in `change_cluster_node_type',
but I want to make sure that these limitations do not have repercussions in
other commands.  We should already take care of the "leave cluster when node
offline" case, but there are probably others...