take into account requeues when setting synch state for slaves
To do this, keep count all the fetches we've seen that require an ack
for messages we don't have (e.g. when the queue we have is shorter than
on the master).

We then decrease this counter appropriately when requeueing, acking, and
set_length'ing.

Given this, we can deem the slave synced only when the length is the same
*and* the counter described above is 9 - there are no pending acks
on the master for messages we don't have.

I might have missed something (I barely tested this) but it seems to do the
trick.