two little fixes...
It turns out that the problems I had with the kill-multi test on default
was because the messages were start publishing right after the queue was
created.  I thought that once queue.declare returned, it meant that the
queue was present on all nodes, but it wasn't, and for this reason we had
the mismatching messages and the purging.

Once I discovered that, I went back two my code and immediately discovered
two very silly mistakes.  I feel silly as well now.

I think that bug 25130 still has a reason to extist, but I wouldn't worry
about it now.