reverse to using `system_info(running_db_nodes)'
We noticed that our function is never necessary.

Doing these changes, I also noticed that `running_db_nodes' does not do an
rpc call if mnesia is running, and instead records the running node in some
ETS table.  This makes the clustering management tests racy, since we don't
know when the information about the nodes which are up/down will propagate.