only trigger stats emission from writers in the server (and not the Erlang client)
The server (and *only* the server) starts channels with
rabbit_channel_sup:start_link({tcp, ...}). Writers are created as part
of that. We pass an extra argument to rabbit_writer:start_link in
this case, indicating that the reader wants be told by the writer to
emit stats.

Internally, the writer handles the stats/no_stats distinction by
initialising the stats timer differently. We introduce a new function
to create a disabled stats timer, thus taking advantage of all the
existing logic in rabbit_event that suppresses timer/stats operations
when stats are disabled.

This approach requires inclusion of rabbit_event in rabbit_common.